ALSA: compress: Remove duplicate ioctl declaration for partial drain
When previous commit was picked, we also had duplicate declaration of
`SNDRV_COMPRESS_PARTIAL_DRAIN`. While this will possibly not generate
any errors or bugs during build and usage, since the value for both
declarations are the same, it will generate a warning that it's
declared multiple times during build. As per Linux kernel guidelines,
we don't want any warnings to generate during build time, so remove
this duplicate.

Fixes: 7d9cef3f4995c5ef65ce3fe05f3792328e9e4de9 (ALSA: compress: Add ioctl declaration for gapless)
Signed-off-by: VoidDev-0 <dev.avax@protonmail.com>
Signed-off-by: Beru Hinode <windowz414@1337.lgbt>
Change-Id: Ibf76389bbc26f230abfe662376ad1f540731816a
Signed-off-by: VoidDev-0 <dev.avax@protonmail.com>