scripts/extract-cert: Fix build for legacy OpenSSL versions
When distros like Gentoo, which would provide an ancient version of
OpenSSL, would compile this kernel, they would hit an error while
doing operations relating to certificates. This is due to any
versions of OpenSSL prior 3.x not implementing
`OpenSSL_add_all_algorithns` function. Let's tie invoking this
function to a simple if-else statement that will see if we have
OpenSSL 3.x or not and handle it accordingly.

[windowz414: Reformat commit message.]

Signed-off-by: Woomymy <github@woomy.ovh>
Signed-off-by: Beru Hinode <windowz414@1337.lgbt>
Change-Id: I77199af7b1d5f79e72fa1b7a4b8cfcee4f1a056c